+++ src/Terrabuild/Core/GraphDef.fs
@@ -13,7 +13,6 @@ type ContaineredShellOperation = {
     Command: string
     Arguments: string
     ErrorLevel: int
-    Stdout: string option
 }
 
 [<RequireQualifiedAccess>]