+++ src/Terrabuild/Helpers/Exec.fs
@@ -344,32 +344,29 @@ let execConsoleArguments (workingDir: string) (command: string) arguments (envs:
 let execConsole (workingDir: string) (command: string) (args: string) (envs: Map<string, string>) =
     execConsoleArguments workingDir command (Arguments.Raw args) envs
 
-let execCaptureTimestampedOutputArguments (workingDir: string) (command: string) arguments (envs: Map<string, string>) (logFile: string) captureStdout =
+let execCaptureTimestampedOutputArguments (workingDir: string) (command: string) arguments (envs: Map<string, string>) (logFile: string) =
     let args = renderArguments arguments
     try
         use logWriter = new StreamWriter(logFile)
-        let stdout = if captureStdout then Some (StringBuilder()) else None
         let writeLock = Lock()
-        let lockWrite (capture: bool) (msg: string | null) =
+        let lockWrite (msg: string | null) =
             match msg with
             | NonNull msg ->
                 lock writeLock (fun () ->
                     logWriter.WriteLine(msg)
-                    if capture then
-                        stdout |> Option.iter (fun output -> output.AppendLine(msg) |> ignore)
                 )
             | _ -> ()
 
         Log.Debug("Running and capturing timestamped output of '{Command}' with arguments '{Args}' in working dir '{WorkingDir}'", command, args, workingDir)
         use proc = createProcess workingDir command arguments envs true
-        proc.OutputDataReceived.Add(fun e -> lockWrite true e.Data)
-        proc.ErrorDataReceived.Add(fun e -> lockWrite false e.Data)
+        proc.OutputDataReceived.Add(fun e -> lockWrite e.Data)
+        proc.ErrorDataReceived.Add(fun e -> lockWrite e.Data)
         proc.BeginOutputReadLine()
         proc.BeginErrorReadLine()
         proc.WaitForExit()
-        proc.ExitCode, (stdout |> Option.map string)
+        proc.ExitCode
     with
         | exn -> forwardExternalError($"Process '{command}' with arguments '{args}' in directory '{workingDir}' failed", exn)
 
-let execCaptureTimestampedOutput (workingDir: string) (command: string) (args: string) (envs: Map<string, string>) (logFile: string) captureStdout =
-    execCaptureTimestampedOutputArguments workingDir command (Arguments.Raw args) envs logFile captureStdout
+let execCaptureTimestampedOutput (workingDir: string) (command: string) (args: string) (envs: Map<string, string>) (logFile: string) =
+    execCaptureTimestampedOutputArguments workingDir command (Arguments.Raw args) envs logFile